+++ changelog.md
@@ -4,6 +4,7 @@
 
 - added rule `habapp_rules.network.wol.Wol` to wake up devices via Wake-on-LAN
 - added rules `habapp_rules.energy.virtual_energy_meter.VirtualEnergyMeterSwitch` and `habapp_rules.energy.virtual_energy_meter.VirtualEnergyMeterNumber` to estimate energy consumption of devices without current / energy measurement
+- Added support for multiple energy items to `habapp_rules.energy.config.monthly_report.EnergyShare`
 
 # Version 7.3.0 - 08.02.2025